feat: Update PUT /tasks/:id route to allow updating task title and description
This commit modifies the PUT /tasks/:id route to enable updating the title and description of a task. The route now accepts a JSON payload with the updated title and/or description fields. If either field is provided, the task's updated_at property is set to the current date and time.

This change enhances the functionality of the application by allowing users to edit the title and description of existing tasks.